+++ cider-mode.el
@@ -59,6 +59,7 @@ entirely."
     (define-key map (kbd "C-c C-d a") 'cider-apropos)
     (define-key map (kbd "C-c C-d A") 'cider-apropos-documentation)
     (define-key map (kbd "C-c C-d g") 'cider-grimoire)
+    (define-key map (kbd "C-c C-d h") 'cider-grimoire-web)
     (define-key map (kbd "C-c C-d j") 'cider-javadoc)
     (define-key map (kbd "C-c C-d d") 'cider-doc)
     (define-key map (kbd "C-M-x")   'cider-eval-defun-at-point)
@@ -122,6 +123,7 @@ entirely."
         ["Display documentation" cider-doc]
         ["Display JavaDoc" cider-javadoc]
         ["Display Grimoire documentation" cider-grimoire]
+        ["Display Grimoire documentation in browser" cider-grimoire-web]
         ["Inspect" cider-inspect]
         "--"
         ["Run test" cider-test-run-test]